+++ src/dicognito/element_anonymizer.py
@@ -10,7 +10,7 @@ class ElementAnonymizer:
     Base class for element anonymizers.
 
     Typically will alter (remove or change) the value of a supplied
-    data_element to obscure patient identifiers. Where a DICOM attribute
+    data_element to obscure patient identifiers. Where a DICOM element
     is linked to another in the same dataset (such as Study Date and Study
     Time), may alter additional elements.
     """